+++ packages/api/test/integration/schema.test.ts
@@ -75,6 +75,7 @@ const QUERIES = [
   'accountProfile',
   'accountName',
   'validateUser',
+  'pickupPoints',
 ]
 
 const MUTATIONS = [