+++ packages/core/src/sdk/product/useLocalizedVariables.ts
@@ -1,7 +1,11 @@
-import type { ClientManyProductsQueryQueryVariables } from '@generated/graphql'
 import { useMemo } from 'react'
+import deepmerge from 'deepmerge'
+
+import { useSearch, type SearchState } from '@faststore/sdk'
+import type { ClientManyProductsQueryQueryVariables } from '@generated/graphql'
+
+import { useSession } from 'src/sdk/session'
 import { ITEMS_PER_SECTION } from 'src/constants'
-import { useSession } from '../session'
 
 const toArray = <T>(x: T[] | T | undefined) =>
   Array.isArray(x) ? x : x ? [x] : []
@@ -17,15 +21,13 @@ export const useLocalizedVariables = ({
   const { channel, locale } = useSession()
 
   return useMemo(() => {
-    const facets = toArray(selectedFacets)
-
     return {
       first: first ?? ITEMS_PER_SECTION,
       after: after ?? '0',
       sort: sort ?? ('score_desc' as const),
       term: term ?? '',
       selectedFacets: [
-        ...facets,
+        ...toArray(selectedFacets),
         { key: 'channel', value: channel ?? '' },
         { key: 'locale', value: locale },
       ],